+++ stdlib/LinearAlgebra/test/generic.jl
@@ -119,7 +119,7 @@ end
     @test [linreg(x,y)...] ≈ [2.5559090909090867, 1.6960139860139862]
     @test [linreg(view(x,1:6),view(y,1:6))...] ≈ [3.8366666666666642,1.3271428571428574]
 
-    # check (LinSpace, UnitRange)
+    # check (LinRange, UnitRange)
     x = range(1.0, stop=12.0, length=100)
     y = -100:-1
     @test [linreg(x, y)...] ≈ [-109.0, 9.0]
@@ -129,7 +129,7 @@ end
     y = 12:-1:1
     @test [linreg(x, y)...] ≈ [13.0, -1.0]
 
-    # check (LinSpace, LinSpace)
+    # check (LinRange, LinRange)
     x = range(-5, stop=10, length=100)
     y = range(50, stop=200, length=100)
     @test [linreg(x, y)...] ≈ [100.0, 10.0]